Add files via upload
This is the first code I wrote with the help of DeepSeek. I tried to add a lot of comments on why each line was written. The first part of the code is the neural network. The second part is just computing the average match length to see that the error obtained in the test is much smaller compared to the average match length. In the code the categorical data was converted into numerical data. But, it is kinda important to keep track of the categorical data to see what are the questions one should come up with to have a successful match(this is one of the questions in the grad division). 

Also, the training was only done using the training.xlx file. I am not entirely sure how to use the file with truncated data and Novice.xlx.