Add files via upload
Sorted the data with months in increasing order down the column corresponding to the same match Id. This is the first code in the python notebook. The second code computes the match length as the difference of completion date and activation date in each row instead of total match length corresponding to a match Id. You can run the code with the xlx file that Will uploaded instead of my csv file.