Add files via upload
This is the first file with recurrent neural network that does sentiment analysis on the contact notes and does a time dependent analysis. To run the code you might have to install transformers or sentence-transformers in your computer. You can run pip3 install transformers and pip3 install sentence-transformers in the terminal. There might be other ways to do it. I don't know about that. Also, the code takes into account the number of contacts and the average match length. It doesn't include any other columns in the data set. So, there is huge room for improvement. Let me know if you have any trouble running the code. The code will take a little bit of time to run by the way.